+++ sdtv2/SDT2PrintMarkdown.py
@@ -2,10 +2,13 @@
 #
 #	Print SDT2 to markdown
 
+from .SDT2Classes import *
 
+hideDetails 				= False
+tables 						= False
+pageBreakBeforeMCandDevices = False
 
-hideDetails = False
-tables = False
+pageBreakToken				= '\n<!--BREAK-->'
 
 # tabulator level
 tab = 0
@@ -53,9 +56,10 @@ def markdownHeader(text):
 #
 
 def print2DomainMarkdown(domain, options):
-	global hideDetails
-	hideDetails = options['hideDetails']
-	tables = options['markdowntables']
+	global hideDetails, pageBreakToken, pageBreakBeforeMCandDevices
+	hideDetails 				= options['hideDetails']
+	tables 						= options['markdowntables']
+	pageBreakBeforeMCandDevices = options['pageBreakBeforeMCandDevices']
 
 	if tables:
 		print('Tables are not supported for input format "sdt2"')
@@ -73,6 +77,8 @@ def print2DomainMarkdown(domain, options):
 		incHeaderLevel()
 		result += markdownHeader('ModuleClasses')
 		for module in domain.modules:
+			if pageBreakBeforeMCandDevices:
+				result += newLine() + pageBreakToken
 			result +=  newLine() + printModuleClass(module)
 		decHeaderLevel()
 
@@ -80,6 +86,8 @@ def print2DomainMarkdown(domain, options):
 		incHeaderLevel()
 		result += markdownHeader('RootDevices')
 		for rootDevice in domain.rootDevices:
+			if pageBreakBeforeMCandDevices:
+				result += newLine() + pageBreakToken
 			result += newLine() + printRootDevice(rootDevice)
 		decHeaderLevel()