+++ sdtv2/__init__.py
@@ -1,0 +1,3 @@
+from sdtv2.SDT2Parser import SDT2Parser
+from sdtv2.SDT2Classes import *
+