+++ sdtv3/SDT3PrintJava.py
@@ -5,7 +5,7 @@
 # TODO export properties
 
 import datetime, os, pathlib, string
-from SDT3Classes import *
+from .SDT3Classes import *
 
 # Dictionary to temporarly store the found structs.
 # TODO: export for each ModuleClass (as done yet) or only once? Check for duplicates?