+++ sdtv3/SDT3PrintOneM2MXSD.py
@@ -1,10 +1,11 @@
-#	SDT3PrintOneM2MSVG.py
+#	SDT3PrintOneM2MXSD.py
 #
 #	Generate XSD for oneM2M
 
 import csv, datetime, os, pathlib, re, string
-from SDT3Classes import *
-import SDTAbbreviate 
+from .SDT3Classes import *
+from common.SDTAbbreviate import *
+from common import SDTAbbreviate
 
 
 # variable that holds an optional header text
@@ -22,15 +23,22 @@ abbreviationsInFile = ''
 # the target name space for the XSD
 targetNamespace = ''
 
+# variable that holds the version of the data model
+modelVersion = ''
+
 # variable for found enum types
 enumTypes = set()
 
 # variable for found actions
 actions = {}
 
+# constants for abbreviation file
+constAbbreviationCSVFile = '_Abbreviations.csv'
+constAbbreviationMAPFile = '_Abbreviations.py'
+
 def print3OneM2MXSD(domain, directory, options):
 	global headerText, domainDefinition, namespacePrefix, enumTypes, doAbbreviations, targetNamespace
-	global abbreviationsInFile
+	global abbreviationsInFile, modelVersion, constAbbreviationCSVFile
 
 	#
 	#	Get various parameters
@@ -49,7 +57,7 @@ def print3OneM2MXSD(domain, directory, options):
 
 	namespacePrefix = options['namespaceprefix']
 	if namespacePrefix == None:			# ERROR
-		print('Error: XSD name space prefix not set')
+		print('Error: name space prefix not set')
 		return
 
 	# get target name space 
@@ -60,19 +68,25 @@ def print3OneM2MXSD(domain, directory, options):
 	# get in and out file names for abbreviations
 	abbreviationsInFile = options['abbreviationsinfile']
 
+	# get the version of the data model
+	modelVersion = options['modelversion']
 
-
-	# Read abbreviations
-	SDTAbbreviate.readAbbreviations(abbreviationsInFile)
 
 	# Create package path and make directories
-	packagePath = directory + os.sep + domain.id.replace('.', os.sep)
-	path = pathlib.Path(packagePath)
+	path = pathlib.Path(directory)
 	try:
 		path.mkdir(parents=True)
 	except FileExistsError as e:
 		pass # on purpose. We override files for now
 	package = sanitizePackage(domain.id)
+
+
+	# Read abbreviations
+	SDTAbbreviate.readAbbreviations(abbreviationsInFile)
+	# Add already existing abbreviations
+	localAbbreviationsfile = str(path) + os.sep + constAbbreviationCSVFile
+	SDTAbbreviate.readAbbreviations(localAbbreviationsfile, predefined=False)
+
 
 	# Export ModuleClasses
 	for module in domain.modules:
@@ -99,15 +113,13 @@ def print3OneM2MXSD(domain, directory, options):
 # Export a ModuleClass definition to a file
 
 def exportModuleClass(module, package, path, name=None):
+	global namespacePrefix, modelVersion
 
 	# export the module class itself
-	prefix = ''
-	if name != None:
-		prefix = sanitizeName(name, False) + '_'
 
 	moduleName   = sanitizeName(module.name, False)
 	fileName     = sanitizeName(module.name, False)
-	fullFilename = str(path) + os.sep + prefix + fileName + '.xsd'
+	fullFilename = getVersionedFilename(fileName, name=name, path=str(path), isModule=True)
 	outputFile   = None
 	try:
 		outputFile = open(fullFilename, 'w')
@@ -124,9 +136,9 @@ def exportModuleClass(module, package, path, name=None):
 def getModuleClassXSD(module, package, name, path):
 
 	result  = ''
-	result += addModuleClassHeader()
+	result += addModuleClassHeader(module)
 	result += addModuleClass(module, name)
-	result += addModuleClassFooter()
+	result += addModuleClassFooter(module)
 	return result
 
 
@@ -141,18 +153,17 @@ xsdSchemaTemplateHeader = '''<?xml version="1.0" encoding="UTF-8"?>
 	xmlns:m2m="http://www.onem2m.org/xml/protocols" xmlns:{namespace}="http://www.onem2m.org/xml/protocols/{domain}" elementFormDefault="unqualified" attributeFormDefault="unqualified"
 	xmlns:xs="http://www.w3.org/2001/XMLSchema">
 
-	<xs:import namespace="http://www.onem2m.org/xml/protocols" schemaLocation="CDT-subscription-v2_5_0.xsd" />
-	<xs:import namespace="http://www.onem2m.org/xml/protocols" schemaLocation="CDT-commonTypes-v2_5_0.xsd" />
+	<xs:import namespace="http://www.onem2m.org/xml/protocols" schemaLocation="CDT-subscription-v2_7_0.xsd" />
+	<xs:import namespace="http://www.onem2m.org/xml/protocols" schemaLocation="CDT-commonTypes-v2_7_0.xsd" />
 
-	<xs:include schemaLocation="CDT-hd_enumerationTypes-v2_5_0.xsd" />
-
+	<xs:include schemaLocation="HD-enumerationTypes-v1_0_0.xsd" />{schemas}
 '''
 
 xsdSchemaTemplateFooter = '''
 </xs:schema>
 '''
 
-def addModuleClassHeader():
+def addModuleClassHeader(module=None):
 	global xsdSchemaTemplateHeader, namespacePrefix
 	global headerText, domainDefinition, targetNamespace
 
@@ -161,10 +172,11 @@ def addModuleClassHeader():
 	return xsdSchemaTemplateHeader.format(headerText=headerText,
 										  namespace=namespacePrefix,
 										  domain=domainDefinition,
-										  targetnamespace=targetNamespace)
+										  targetnamespace=targetNamespace,
+										  schemas=getModuleClassSchemas(module))
 
 
-def addModuleClassFooter():
+def addModuleClassFooter(module=None):
 	global xsdSchemaTemplateFooter
 	decTab()
 	return xsdSchemaTemplateFooter.format()
@@ -173,77 +185,89 @@ def addModuleClassFooter():
 # The main resource definition
 
 flexContainerResourceTemplate = '''
-	<xs:element name="{name}">
-		<xs:complexType>
-			<xs:complexContent>
-				<!-- Inherit Common Attributes from data type "flexContainerResource" -->
-				<xs:extension base="m2m:flexContainerResource">
-					<xs:sequence>
-					
-						<!-- Resource Specific Attributes -->
+	<xs:element name="{name}" type="{namespacePrefix}:{name}" substitutionGroup="m2m:sg_flexContainerResource"/>
+	<xs:complexType name="{name}">
+		<xs:complexContent>
+			<!-- Inherit Common Attributes from data type "flexContainerResource" -->
+			<xs:extension base="m2m:flexContainerResource">
+				<xs:sequence>
+				
+					<!-- Resource Specific Attributes -->
 {specificAttributes}
 
-						<!-- Child Resources -->
+					<!-- Child Resources -->
 
-						<xs:choice minOccurs="0" maxOccurs="1">
-							<xs:element name="childResource" type="m2m:childResourceRef" minOccurs="1" maxOccurs="unbounded" />
-							<xs:choice minOccurs="1" maxOccurs="unbounded">
+					<xs:choice minOccurs="0" maxOccurs="1">
+						<xs:element name="childResource" type="m2m:childResourceRef" minOccurs="1" maxOccurs="unbounded" />
+						<xs:choice minOccurs="1" maxOccurs="unbounded">
 {actionElements}
 
-								<xs:element ref="m2m:subscription"  />
-							</xs:choice>
+							<xs:element ref="m2m:subscription"  />
 						</xs:choice>
+					</xs:choice>
 
-					</xs:sequence>
-				</xs:extension>
-			</xs:complexContent>
-		</xs:complexType>
-	</xs:element>
+				</xs:sequence>
+			</xs:extension>
+		</xs:complexContent>
+	</xs:complexType>
 
-	<xs:element name="{name}Annc">
-		<xs:complexType>
-			<xs:complexContent>
-				<!-- Inherit Common Attributes from data type "announcedFlexContainerResource" -->
-				<xs:extension base="m2m:announcedFlexContainerResource">
-					<xs:sequence>
+	<xs:element name="{name}Annc" type="{namespacePrefix}:{name}Annc" substitutionGroup="m2m:sg_announcedFlexContainerResource"/>
+	<xs:complexType name="{name}Annc">
+		<xs:complexContent>
+			<!-- Inherit Common Attributes from data type "announcedFlexContainerResource" -->
+			<xs:extension base="m2m:announcedFlexContainerResource">
+				<xs:sequence>
 
-						<!-- Resource Specific Attributes -->		
+					<!-- Resource Specific Attributes -->		
 {specificAttributesAnnc}
 
-						<!-- Child Resources -->
+					<!-- Child Resources -->
 
-						<xs:choice minOccurs="0" maxOccurs="1">
-							<xs:element name="childResource" type="m2m:childResourceRef" minOccurs="1" maxOccurs="unbounded" />
-							<xs:choice minOccurs="1" maxOccurs="unbounded">
+					<xs:choice minOccurs="0" maxOccurs="1">
+						<xs:element name="childResource" type="m2m:childResourceRef" minOccurs="1" maxOccurs="unbounded" />
+						<xs:choice minOccurs="1" maxOccurs="unbounded">
 {actionElements}
 
-								<xs:element ref="m2m:subscription"  />
-							</xs:choice>
+							<xs:element ref="m2m:subscription"  />
 						</xs:choice>
+					</xs:choice>
 
-					</xs:sequence>
-				</xs:extension>
-			</xs:complexContent>
-		</xs:complexType>
-	</xs:element>
+				</xs:sequence>
+			</xs:extension>
+		</xs:complexContent>
+	</xs:complexType>
 '''
 
 def addModuleClass(module, name):
 	global flexContainerResourceTemplate, namespacePrefix
 	name = sanitizeName(name, False)
-	nameAnnc = sanitizeName(name + 'Annc', False)
-	return flexContainerResourceTemplate.format(name=name, 
-												namespace=namespacePrefix,
-												specificAttributes=getSpecificAttributes(module, annc=False),
-												specificAttributesAnnc=getSpecificAttributes(module, annc=True),
-												actionElements=getSpecificActions(module))
+	result = flexContainerResourceTemplate.format(name=name,
+												  namespacePrefix=namespacePrefix,
+												  specificAttributes=getSpecificAttributes(module, annc=False),
+						 						  specificAttributesAnnc=getSpecificAttributes(module, annc=True),
+												  actionElements=getSpecificActions(module))
+
+	# small hack: add the announced device name to the abbreviated list
+	sanitizeName(name, False, annc=True)
+
+	return result
+
 
 
 def getSpecificAttributes(module, annc=False):
 	result = ''
 	incTab(5)
+
+	# Properties
+	if len(module.properties) > 0:
+		for prop in module.properties:
+			result += getSpecificPropertyXSD(prop, annc)
+		result += newLine()
+
+	# DataPoints
 	for data in module.data:
 		result += getDataPointXSD(data, annc)
+
 	decTab(5)
 	return result
 
@@ -275,17 +299,21 @@ def getDataPointXSD(data, annc):
 
 
 def getDataPointType(dataPoint):
+	name = sanitizeName(dataPoint.name, False)
+	return getDataType(name, dataPoint.type)
+
+
+def getDataType(name, typ):
 	global enumTypes, namespacePrefix
 	result = ''
-	name = sanitizeName(dataPoint.name, False)
 
 	# Array
-	if (isinstance(dataPoint.type.type, SDT3ArrayType)):
+	if (isinstance(typ.type, SDT3ArrayType)):
 		result += '>'
 		incTab()
 		result += newLine() + '<xs:simpleType>'
 		incTab()
-		result += newLine() + '<xs:list itemType="' + namespacePrefix + ':' + name + '" />'
+		result += newLine() + '<xs:list itemType="' + getSimpleDataType(typ.type.arrayType.type.type) + '" />'
 		decTab()
 		result += newLine() + '</xs:simpleType>'
 		decTab()
@@ -293,8 +321,8 @@ def getDataPointType(dataPoint):
 
 
 	# Simple type
-	if isinstance(dataPoint.type.type, SDT3SimpleType):
-		ty = dataPoint.type.type
+	if isinstance(typ.type, SDT3SimpleType) or isinstance(typ, SDT3SimpleType):
+		ty = typ.type if isinstance(typ.type, SDT3SimpleType) else typ
 		if ty.type == 'enum':
 			result += ' >'
 			incTab()
@@ -305,18 +333,19 @@ def getDataPointType(dataPoint):
 			result += newLine() + '</xs:simpleType>'
 			decTab()
 			result += newLine() + '</xs:element>'
-			enumTypes.add(dataPoint.name)
+			enumTypes.add(name)
 		else:
 			result += ' type="'
 			result += getSimpleDataType(ty.type)
 			result += '" />'
 
 	# Struct
-	elif (isinstance(dataPoint.type.type, SDT3StructType)):
+	elif (isinstance(typ.type, SDT3StructType)):
 		print('structType not supported yet')
 		result += 'XXX'
 	
 	return result
+
 
 def getSimpleDataType(type):
 	result = ''
@@ -335,7 +364,7 @@ def getSimpleDataType(type):
 	elif type == 'time':
 		result += 'm2m:timestamp'	# CHECK
 	elif type =='uri':
-		result += 'm2m:URI'			# CHECK
+		result += 'xs:anyURI'
 	elif type == 'blob':
 		result += 'xs:base64Binary'	# CHECK 
 	elif re.match('.+:.+', type):	# CHECK enum
@@ -398,7 +427,7 @@ def getEnumType(enumName):
 	result  = xsdSchemaTemplateHeader.format(headerText=headerText, namespace=namespacePrefix, domain=domainDefinition, targetnamespace=targetNamespace)
 	result += xsdSchemaTemplateEnumHeader.format(name=sanitizeName(enumName, False))
 	incTab(3)
-	result += newLine() + '<!-- TODO comment -->'
+	#result += newLine() + '<!-- TODO comment -->'
 	result += newLine() + '<xs:enumeration value="1" />'
 	decTab(3)
 	result += xsdSchemaTemplateEnumFooter.format()
@@ -410,7 +439,7 @@ def getEnumType(enumName):
 # Actions
 
 xsdActionTemplate = '''
-	<xs:element name="{name}">
+	<xs:element name="{name}" substitutionGroup="m2m:sg_flexContainerResource">
 		<xs:complexType>
 			<xs:complexContent>
 				<!-- Inherit Common Attributes from data type "flexContainerResource" -->
@@ -433,7 +462,7 @@ xsdActionTemplate = '''
 		</xs:complexType>
 	</xs:element>
 
-	<xs:element name="{name}Annc">
+	<xs:element name="{name}Annc" substitutionGroup="m2m:sg_announcedFlexContainerResource">
 		<xs:complexType>
 			<xs:complexContent>
 				<!-- Inherit Common Attributes from data type "announcedFlexContainerResource" -->
@@ -465,8 +494,8 @@ def exportActions(path):
 
 
 def exportAction(path, action):
-	fileName = sanitizeName(action.name, False)
-	fullFileName = str(path) + os.sep + 'Action_' + fileName + '.xsd'
+	name = sanitizeName(action.name, False)
+	fullFileName = getVersionedFilename(name, path=str(path), isAction=True)
 	outputFile = None
 	try:
 		outputFile = open(fullFileName, 'w')
@@ -485,7 +514,27 @@ def getAction(action):
 	result += addModuleClassHeader()
 	result += xsdActionTemplate.format(name=sanitizeName(action.name, False))
 	result += addModuleClassFooter()
+
+	# small hack: add the announced device name to the abbreviated list
+	sanitizeName(action.name, False, annc=True)
+
 	return result
+
+
+#############################################################################
+
+# Properties
+
+def getSpecificPropertyXSD(prop, annc):
+	result = ''
+	result += newLine() + '<xs:element name="' + sanitizeName(prop.name, False) + '"'
+	if annc:
+		result += ' minOccurs="0"'
+	else:
+		if prop.optional == 'true':	# indicate an optional property
+			result += ' minOccurs="0"'
+	result += getDataType(prop.name, prop.type)
+	return result;
 
 
 #############################################################################
@@ -501,10 +550,10 @@ xsdSchemaTemplateDeviceHeader = '''<?xml version="1.0" encoding="UTF-8"?>
 	xmlns:m2m="http://www.onem2m.org/xml/protocols" xmlns:{namespace}="http://www.onem2m.org/xml/protocols/{domain}" elementFormDefault="unqualified" attributeFormDefault="unqualified"
 	xmlns:xs="http://www.w3.org/2001/XMLSchema">
 
-	<xs:import namespace="http://www.onem2m.org/xml/protocols" schemaLocation="CDT-subscription-v2_5_0.xsd" />
-	<xs:import namespace="http://www.onem2m.org/xml/protocols" schemaLocation="CDT-commonTypes-v2_5_0.xsd" />
+	<xs:import namespace="http://www.onem2m.org/xml/protocols" schemaLocation="CDT-subscription-v2_7_0.xsd" />
+	<xs:import namespace="http://www.onem2m.org/xml/protocols" schemaLocation="CDT-commonTypes-v2_7_0.xsd" />
 
-	<xs:include schemaLocation="CDT-hd_enumerationTypes-v2_5_0.xsd" />{moduleSchemas}
+	<xs:include schemaLocation="HD-enumerationTypes-v1_0_0.xsd" />{schemas}
 
 '''
 
@@ -513,7 +562,7 @@ xsdSchemaTemplateDeviceFooter = '''
 '''
 
 flexContainerDeviceResourceTemplate = '''
-	<xs:element name="{deviceName}">
+	<xs:element name="{deviceName}" substitutionGroup="m2m:sg_flexContainerResource">
 		<xs:complexType>
 			<xs:complexContent>
 				<!-- Inherit Common Attributes from data type "flexContainerResource" -->
@@ -521,7 +570,7 @@ flexContainerDeviceResourceTemplate = '''
 					<xs:sequence>
 					
 						<!-- Resource Specific Attributes -->
-{properties}
+{deviceProperties}						
 
 						<!-- Child Resources -->
 						
@@ -529,6 +578,7 @@ flexContainerDeviceResourceTemplate = '''
 							<xs:element name="childResource" type="m2m:childResourceRef" minOccurs="1" maxOccurs="unbounded" />
 							<xs:choice minOccurs="1" maxOccurs="unbounded">
 {moduleClasses}
+
 								<xs:element ref="m2m:subscription"  />
 							</xs:choice>
 						</xs:choice>
@@ -539,7 +589,7 @@ flexContainerDeviceResourceTemplate = '''
 		</xs:complexType>
 	</xs:element>
 
-	<xs:element name="{deviceName}Annc">
+	<xs:element name="{deviceName}Annc" substitutionGroup="m2m:sg_announcedFlexContainerResource">
 		<xs:complexType>
 			<xs:complexContent>
 				<!-- Inherit Common Attributes from data type "announcedFlexContainerResource" -->
@@ -547,14 +597,14 @@ flexContainerDeviceResourceTemplate = '''
 					<xs:sequence>
 
 						<!-- Resource Specific Attributes -->		
-{properties}					
-
+{devicePropertiesAnnc}
 						<!-- Child Resources -->
 
 						<xs:choice minOccurs="0" maxOccurs="1">
 							<xs:element name="childResource" type="m2m:childResourceRef" minOccurs="1" maxOccurs="unbounded" />
 							<xs:choice minOccurs="1" maxOccurs="unbounded">
-{moduleClasses}
+{moduleClassesAnnc}
+
 								<xs:element ref="m2m:subscription"  />
 							</xs:choice>
 						</xs:choice>
@@ -568,22 +618,17 @@ flexContainerDeviceResourceTemplate = '''
 
 
 def exportDevice(device, package, path):
-	deviceName   = sanitizeName(device.id, False)
-	fileName     = sanitizeName(device.id, False)
-	packagePath  = str(path) + os.sep + deviceName.lower()
+	global namespacePrefix, modelVersion
 
-	# create sub-directory fist
-	path = pathlib.Path(packagePath)
-	try:
-		path.mkdir(parents=True)
-	except FileExistsError as e:
-		pass # on purpose. We override files for now
+	# export the module class itself
 
-	fullFilename = str(path) + os.sep + fileName + '.xsd'
+	name 	     = sanitizeName(device.id, False)
+	fullFilename = getVersionedFilename(name, name=None, path=str(path))
 	outputFile   = None
+
 	try:
 		outputFile = open(fullFilename, 'w')
-		outputFile.write(getDeviceXSD(device, package, deviceName, path))		
+		outputFile.write(getDeviceXSD(device, package, name, path))		
 	except IOError as err:
 		print(err)
 	finally:
@@ -593,9 +638,11 @@ def exportDevice(device, package, path):
 
 def getDeviceXSD(device, package, deviceName, path):
 	result  = ''
+	incTab()
 	result += addDeviceHeader(device)
 	result += addDevice(device, deviceName)
 	result += addDeviceFooter()
+	decTab()
 	return result
 
 
@@ -604,17 +651,15 @@ def addDeviceHeader(device):
 	global headerText, domainDefinition, targetNamespace
 
 	# XML header + license text
-	incTab()
 	return xsdSchemaTemplateDeviceHeader.format(headerText=headerText, 
 												namespace=namespacePrefix,
 												domain=domainDefinition,
 												targetnamespace=targetNamespace,
-												moduleSchemas=getDeviceModuleClassSchemas(device))
+												schemas=getDeviceSchemas(device))
 
 
 def addDeviceFooter():
 	global xsdSchemaTemplateDeviceFooter
-	decTab()
 	return xsdSchemaTemplateDeviceFooter.format()
 
 
@@ -622,36 +667,96 @@ def addDevice(device, deviceName):
 	global flexContainerDeviceResourceTemplate
 	incTab(5)
 	result = flexContainerDeviceResourceTemplate.format(deviceName=deviceName, 
-													    properties=getDeviceProperties(device),
-													    moduleClasses=getDeviceModuleClasses(device))
+														deviceProperties=getDeviceProperties(device),
+														devicePropertiesAnnc=getDeviceProperties(device, annc=True),
+													    moduleClasses=getDeviceModuleClasses(device),
+													    moduleClassesAnnc=getDeviceModuleClasses(device, withAnnc=True))
+
+	# small hack: add the announced device name to the abbreviated list
+	sanitizeName(deviceName, False, annc=True)
+
 	decTab(5)
 	return result
 
 
-def getDeviceModuleClassSchemas(device):
+def getDeviceSchemas(device):
+	global namespacePrefix
+
 	result = ''
+
+	# Referenced modules
+	parentModuleClasses = {}
+	parentModuleClassNames = []
 	for module in device.modules:
-		result += newLine() + '<xs:include schemaLocation="' + sanitizeName(module.name, False) + '.xsd" />'
+		parentModuleClassName = module.extends.domain + '.' + module.extends.clazz
+		if module.name == module.extends.clazz:
+			name = getVersionedFilename(module.name, isModule=True)
+			result += newLine() + '<xs:include schemaLocation="' + name + '" />'
+		else:
+			parentModuleClasses[module.name] = module
+			if parentModuleClassName not in parentModuleClassNames:
+				name = getVersionedFilename(module.extends.clazz, isModule=True)
+				result += newLine() + '<xs:include schemaLocation="' + name + '" />'
+				parentModuleClassNames.append(parentModuleClassName)
+
+	# Special handling when the extended class is actually different from the parent class 
+	extraElements = ''
+	if len(parentModuleClasses) > 0:
+		extraElements += newLine()
+		for name in parentModuleClasses:
+			module = parentModuleClasses[name]
+			# For now, we only check for a different name. Here we need to create a new XSD but with the same content as the orignal class.
+			if module.name != module.extends.clazz:
+				extraElements += newLine() + '<xs:element name="' + module.name + '" type="' + namespacePrefix + ':' + module.extends.clazz + '" />'
+				extraElements += newLine() + '<xs:element name="' + module.name + 'Annc" type="' + namespacePrefix + ':' + module.extends.clazz + 'Annc" />'
+
+			#	Future improvement to add more elements when extending the inherited ModuleClass
+			#   <xs:element name="freshTemperature">
+            #		<xs:complexType>
+			#			<xs:complexContent>
+			#				<xs:extension base="hd:temperature">
+			#
+			#				</xs:extension>
+			#			</xs:complexContent>
+			#		</xs:complexType>
+			#	</xs:element>
+
+	result += extraElements
 	return result
 
 
-def getDeviceProperties(device):
-	global namespacePrefix
+def getModuleClassSchemas(module):
+	if module == None:
+		return ''
 	result = ''
-	for prop in device.properties:
-		optional = '0' if prop.optional == 'true' else '1'
-		result += newLine() + '<xs:element name="' + namespacePrefix + ':' + sanitizeName(prop.name, False) + '" minOccurs="' + optional + '" type="xs:string" />'
-	return result;
+
+	# Referenced Actions
+	for action in module.actions:
+		name = getVersionedFilename(action.name, isAction=True)
+		result += newLine() + '<xs:include schemaLocation="' + name + '" />'
+	return result
 
 
-def getDeviceModuleClasses(device):
+def getDeviceModuleClasses(device, withAnnc=False):
 	global namespacePrefix
 	result = ''
 	incTab(2)
 	for module in device.modules:
-		result += newLine() + '<xs:element ref="' + namespacePrefix + ':' + sanitizeName(module.name, False) + '" />'
+		name = sanitizeName(module.name, False, annc=withAnnc) # Side effect: add abbreviated annc resource
+		result += newLine() + '<xs:element ref="' + namespacePrefix + ':' + name + '" />'
+		# if this is an announced ressource, then add an extra line for it
+		if withAnnc:
+			result += newLine() + '<xs:element ref="' + namespacePrefix + ':' + name + 'Annc" />'
 	decTab(2)
 	return result;
+
+
+def getDeviceProperties(device, annc=False):
+	result = ''
+	# Properties
+	for prop in device.properties:
+		result += getSpecificPropertyXSD(prop, annc)
+	return result
 
 
 
@@ -660,9 +765,10 @@ def getDeviceModuleClasses(device):
 # Export abbreviations
 
 def exportAbbreviations(path, abbreviations):
+	global constAbbreviationCSVFile, constAbbreviationMAPFile
 
 	# Export as python map
-	fullFilename = str(path) + os.sep + '_Abbreviations.py'
+	fullFilename = str(path) + os.sep + constAbbreviationMAPFile
 	outputFile = None
 	try:
 		outputFile = open(fullFilename, 'w')
@@ -674,7 +780,7 @@ def exportAbbreviations(path, abbreviations):
 			outputFile.close()
 
 	# Export as CSV
-	fullFilename = str(path) + os.sep + '_Abbreviations.csv'
+	fullFilename = str(path) + os.sep + constAbbreviationCSVFile
 	outputFile = None
 	try:
 		outputFile = open(fullFilename, 'w', newline='')
@@ -714,7 +820,7 @@ def getAbbreviations(abbreviations):
 
 # Sanitize the name for SVG
 
-def sanitizeName(name, isClass):
+def sanitizeName(name, isClass, annc=False):
 	if (name == None or len(name) == 0):
 		return ''
 	result = name
@@ -733,16 +839,49 @@ def sanitizeName(name, isClass):
 	result = result.replace(')', '_')
 	result = result.replace('-', '_')
 
-	SDTAbbreviate.abbreviate(result)
-	
+	# If this name is an announced resource, add "Annc" Postfix to both the
+	# name as well as the abbreviation.
+	if annc:
+		abbr = SDTAbbreviate.abbreviate(result)
+		SDTAbbreviate.addAbbreviation(result + 'Annc', abbr + 'Annc')
+	else:
+		abbr = SDTAbbreviate.abbreviate(result)
+		SDTAbbreviate.addAbbreviation(result, abbr)
+
 	return result
 
-# Sanitize the package name for SVG
+# Sanitize the package name for XSD
 
 def sanitizePackage(package):
 	result = package.replace('/', '.')
 	return result
 
+# get a versioned filename
+
+def getVersionedFilename(fileName, name=None, path=None, isModule=False, isAction=False):
+	global modelVersion, namespacePrefix
+
+	prefix  = ''
+	postfix = ''
+	if name != None:
+		prefix += sanitizeName(name, False) + '_'
+	else:
+		if namespacePrefix != None:
+			prefix += namespacePrefix.upper() + '-'
+		if isAction:
+			prefix += 'act-'
+		if isModule:
+			prefix += 'mod-'
+
+	if modelVersion != None:
+		postfix += '-v' + modelVersion.replace('.', '_')
+
+	fullFilename = ''
+	if path != None:
+		fullFilename = path + os.sep
+	fullFilename += prefix + sanitizeName(fileName, False) + postfix + '.xsd'
+
+	return fullFilename
 
 
 # Tabulator handling