+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 
-## Version 0.7
+## Version 0.8
 xx.xx.xxxx
+
+- Added *--modelversion** 
+- Adopted to new naming scheme for oneM2M XSD and SVG files
+- Various fixes for XSD schemas. Support oneM2M Version 2 XSD
+- Refactored script directory structure
+- Added *--markdownpagebreak* option
+ 
+
+## Version 0.7
+09.04.2016
 
 - Export to Eclipse Vorto, first version
 - First version of export to SVG in oneM2M resource format